// moonfowcl/src/app/model/common/consumer-params.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */

export class ConsumerParams {
  consumer: string;
  service: string;
  codiceModulo: string;
  codiceEnte: string;
  codiceIstanza: string;
  backUrl: string;
  parameters: any;
  faq: string;
  assistenza: string;
}